Add files via upload
The potential of customized music recommendation systems to boost user happiness and engagement has gained a lot of attention in recent years. The use of non-verbal clues like facial expressions to provide better suggestions that take the user’s situation into account is gaining popularity. In this study, we provide an innovative approach to music recommendation using facial expressions, with a focus on creating unique listening sessions for each individual user.